+++ app/models/rails_doc_generator/user.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module RailsDocGenerator
+  class User < ApplicationRecord
+    self.table_name = 'rails_doc_generator_users'
+    
+    has_secure_password
+    
+    validates :username, presence: true, uniqueness: true
+    validates :password, length: { minimum: 6 }, if: -> { password.present? }
+  end
+end