+++ lib/rails_doc_generator/configuration.rb
@@ -3,7 +3,7 @@
 module RailsDocGenerator
   class Configuration
     attr_accessor :output_dir, :app_name, :include_timestamps, :include_validations,
-                  :include_scopes, :theme_color
+                  :include_scopes, :theme_color, :authenticate_with
 
     def initialize
       @output_dir = default_output_dir